# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Na Pista API",
    "version": "1.0.0",
    "description": "Na Pista — operational infrastructure for businesses (catalogue, stock, orders, services,\nprofessionals, scheduling, appointments), part of the Última Linha ecosystem.\n\n**Authentication:** `Authorization: Bearer <token>` — either a UL Platform user access token (Supabase Auth JWT)\nor a UL Platform service credential (`ulk_…`). See docs/api/authentication.md.\n\n**Tenancy:** every business route is under `/organizations/{organizationId}`; the caller must be an active member\n(human) or hold a credential issued for that organization (service).\n\n**Envelope:** success `{ \"data\": … }` (paginated lists add `pagination`); error\n`{ \"error\": { \"code\", \"message\", \"details\"? } }`. Clients must ignore unknown response fields.\n\n**Authorization extensions:** `x-na-pista-permission` (human role permission), `x-na-pista-scope` (service\ncredential scope; absent = humans only), `x-na-pista-capability` (organization entitlement required)."
  },
  "servers": [
    {
      "url": "/v1",
      "description": "Relative to the Na Pista API host."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/organizations/{organizationId}/categories": {
      "post": {
        "operationId": "postOrganizationsByOrganizationIdCategories",
        "tags": [
          "Categories"
        ],
        "summary": "Create a category",
        "x-na-pista-permission": "categories.create",
        "x-na-pista-scope": "catalog.write",
        "x-na-pista-capability": "catalog.enabled",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 2000
                  }
                },
                "required": [
                  "name"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Category"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (see `error.details`). Codes: `VALIDATION_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or rejected credentials. Codes: `UNAUTHORIZED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of / not scoped to this organization, missing permission or scope, or capability not enabled. Codes: `FORBIDDEN`, `ENTITLEMENT_REQUIRED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state. Codes: `CONFLICT`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body too large. Codes: `PAYLOAD_TOO_LARGE`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error. Report the X-Request-ID. Codes: `INTERNAL_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A required dependency (UL Platform or Na Pista's database) is unavailable. Retry later. Codes: `UPSTREAM_UNAVAILABLE`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getOrganizationsByOrganizationIdCategories",
        "tags": [
          "Categories"
        ],
        "summary": "List categories",
        "x-na-pista-permission": "categories.read",
        "x-na-pista-scope": "catalog.read",
        "x-na-pista-capability": "catalog.enabled",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "ARCHIVED"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "default": 50,
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Category"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (see `error.details`). Codes: `VALIDATION_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or rejected credentials. Codes: `UNAUTHORIZED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of / not scoped to this organization, missing permission or scope, or capability not enabled. Codes: `FORBIDDEN`, `ENTITLEMENT_REQUIRED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error. Report the X-Request-ID. Codes: `INTERNAL_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A required dependency (UL Platform or Na Pista's database) is unavailable. Retry later. Codes: `UPSTREAM_UNAVAILABLE`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{organizationId}/categories/{categoryId}": {
      "get": {
        "operationId": "getOrganizationsByOrganizationIdCategoriesByCategoryId",
        "tags": [
          "Categories"
        ],
        "summary": "Get a category",
        "x-na-pista-permission": "categories.read",
        "x-na-pista-scope": "catalog.read",
        "x-na-pista-capability": "catalog.enabled",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Category"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (see `error.details`). Codes: `VALIDATION_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or rejected credentials. Codes: `UNAUTHORIZED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of / not scoped to this organization, missing permission or scope, or capability not enabled. Codes: `FORBIDDEN`, `ENTITLEMENT_REQUIRED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found in this organization. Codes: `NOT_FOUND`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error. Report the X-Request-ID. Codes: `INTERNAL_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A required dependency (UL Platform or Na Pista's database) is unavailable. Retry later. Codes: `UPSTREAM_UNAVAILABLE`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "patchOrganizationsByOrganizationIdCategoriesByCategoryId",
        "tags": [
          "Categories"
        ],
        "summary": "Update a category",
        "x-na-pista-permission": "categories.update",
        "x-na-pista-scope": "catalog.write",
        "x-na-pista-capability": "catalog.enabled",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "description": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 2000
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "ARCHIVED"
                    ]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Category"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (see `error.details`). Codes: `VALIDATION_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or rejected credentials. Codes: `UNAUTHORIZED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of / not scoped to this organization, missing permission or scope, or capability not enabled. Codes: `FORBIDDEN`, `ENTITLEMENT_REQUIRED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found in this organization. Codes: `NOT_FOUND`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state. Codes: `CONFLICT`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body too large. Codes: `PAYLOAD_TOO_LARGE`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error. Report the X-Request-ID. Codes: `INTERNAL_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A required dependency (UL Platform or Na Pista's database) is unavailable. Retry later. Codes: `UPSTREAM_UNAVAILABLE`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteOrganizationsByOrganizationIdCategoriesByCategoryId",
        "tags": [
          "Categories"
        ],
        "summary": "Archive a category",
        "description": "Archives (never physically deletes) and returns the archived category.",
        "x-na-pista-permission": "categories.delete",
        "x-na-pista-scope": "catalog.write",
        "x-na-pista-capability": "catalog.enabled",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Category"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (see `error.details`). Codes: `VALIDATION_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or rejected credentials. Codes: `UNAUTHORIZED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of / not scoped to this organization, missing permission or scope, or capability not enabled. Codes: `FORBIDDEN`, `ENTITLEMENT_REQUIRED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found in this organization. Codes: `NOT_FOUND`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state. Codes: `CONFLICT`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error. Report the X-Request-ID. Codes: `INTERNAL_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A required dependency (UL Platform or Na Pista's database) is unavailable. Retry later. Codes: `UPSTREAM_UNAVAILABLE`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A UL Platform user access token (JWT) or a UL Platform service credential (`ulk_…`)."
      }
    },
    "schemas": {
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "ARCHIVED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 date-time (UTC)."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 date-time (UTC)."
          }
        },
        "required": [
          "id",
          "organizationId",
          "name",
          "description",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable machine-readable code. Branch on this, then on the HTTP status."
              },
              "message": {
                "type": "string",
                "description": "Human-readable; not part of the contract (may change)."
              },
              "details": {
                "description": "Present on VALIDATION_ERROR when field-level information exists.",
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ValidationIssue"
                }
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "ValidationIssue": {
        "type": "object",
        "properties": {
          "location": {
            "type": "string",
            "enum": [
              "body",
              "query",
              "path"
            ]
          },
          "path": {
            "type": "string",
            "description": "Dotted path of the invalid field in the client's input (empty = the whole body)."
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "message"
        ]
      }
    }
  }
}